ASoC: img: rename pistachio card driver files
Rename pistachio card driver files following rewrite
